{{cookiecutter.service_repository}}/requirements/local.txt
-r production.txt
factory-boy
ipdb
pre-commit
pytest-cache
pytest-cov
pytest-django
pytest-timeout
pytest-xdist
sphinx
webtest
vcrpy
